Update firmware to not blink
Got tired of the code sending empty keybuffers when you pressed a layer
key. They should not trigger any sending of keys at all. They are
internal things to the keyboard and should not trigger us to send a
datagram to the host.

This introduced a "fun" bug in the beginning where it would not clear
keys if you first help SHIFT, pressed a layer, released SHIFT, released
layer. It would think that SHIFT was still pressed. This was due to the
new way we check if we should send a datagram. We don't check if the
state has changed anymore for this, we check if the keybuffer has
changed. Solved it by actually checking the state later to see if the
keybuffer should be reset. Solved it and we don't send empty datagrams
anymore. So this solves that but not in the most elegant solution I
would have liked.